- DI using Dip - move entirely to CleanSwift - some updates & fixes

+++ src/Domain/Wallet/WalletViewModel.swift
@@ -11,11 +11,7 @@ import SHTokenField
 import Database
 import Shared
 
-public protocol WalletPresenter: ReloadablePresenter {
-    func reloadWallets()
-}
-
-public final class WalletViewModel: SheklyViewModel {
+public final class WalletViewModel: ViewModel {
     // MARK: - Internal properties
     var selectedMonthDate: Date?
     var selectedWallet: SheklyWalletModel?
@@ -31,14 +27,11 @@ public final class WalletViewModel: SheklyViewModel {
     private let userProvider: UserManaging
     
     // MARK: - Constructor
-    init(
-        presenter: WalletPresenter,
-        dataController: SheklyDataController,
-        differ: Differ,
-        currencyFormatter: SheklyCurrencyFormatter,
-        userProvider: UserManaging
-        ) {
-        
+    init(presenter: WalletPresenter,
+         dataController: SheklyDataController,
+         differ: Differ,
+         currencyFormatter: SheklyCurrencyFormatter,
+         userProvider: UserManaging) {
         self.presenter = presenter
         self.dataController = dataController
         self.differ = differ
@@ -50,12 +43,10 @@ public final class WalletViewModel: SheklyViewModel {
         let wallets = dataController.getWallets().map(SheklyWalletModel.init)
         let selectedWallet = wallets.filter { $0.id == userProvider.selectedWalletId }.first
         self.selectedWallet = selectedWallet ?? wallets.first
-        
-        super.init()
     }
     
     // MARK: - Public methods
-    public override func viewDidAppear() {
+    public func viewWillAppear() {
         reloadEntries()
         reloadWallets()
     }
@@ -89,23 +80,21 @@ public extension WalletViewModel {
     }
     
     func numberOfEntries() -> Int {
-        let count = entries.count
-        guard count > 0 else {
-            return 1
-        }
-        
-        return count
+        return entries.count
     }
     
-    func entryModel(forIndexPath indexPath: IndexPath) -> SheklyEntryModel {
-        let model: SheklyEntryModel? = entries[safe: indexPath.row]
+    func entryModel(forIndexPath indexPath: IndexPath) -> SheklyWalletEntryModel {
+        let model: SheklyWalletEntryModel? = entries[safe: indexPath.row]
         
         return model ?? SheklyEntryEmptyModel()
     }
     
     func deleteEntry(atIndexPath indexPath: IndexPath) -> Bool {
-        let entries: [WalletEntryModel] = self.entries.map { $0.entry }
-        let entry: WalletEntryModel = entries[indexPath.row]
+        let entries: [WalletEntryModel] = self.entries.compactMap { $0.entry }
+        
+        guard let entry = entries[safe: indexPath.row] else {
+            return false
+        }
         
         let success = dataController.delete(entry: entry)
         
@@ -147,40 +136,26 @@ extension WalletViewModel {
                 return SheklyWalletEntryModel(entry: entry, formatter: currencyFormatter)
         }
         
-        let models: [SheklyEntryModel]
-            
-        if entries.isEmpty == true {
-            models = [SheklyEntryEmptyModel()]
-        } else {
-            models = entryModels
-        }
-        
-        let oldState: [SheklyEntryModel]
-        
-        if self.entries.isEmpty == true {
-            oldState = [SheklyEntryEmptyModel()]
-        } else {
-            oldState = self.entries
-        }
-        
-        self.entries = entryModels
+        let models: [SheklyWalletEntryModel] = entries.isEmpty ? [SheklyEntryEmptyModel()] : entryModels
+        let oldState = self.entries
         
         let changeSet: ChangeSet = differ.getDiff(oldState: oldState, newState: models)
         
-        presenter?.reload(changeSet: changeSet)
+        presenter?.reload(changeSet: changeSet, setData: { [weak self] in
+            self?.entries = models
+        })
     }
 }
 
 private extension Array where Element == WalletEntryModel {
     func sorted() -> [Element] {
-        return self
-            .sorted { (left, right) -> Bool in
-                guard let leftDate = left.date,
-                    let rightDate = right.date else {
-                        return false
-                }
-                
-                return leftDate > rightDate
+        return sorted { (left, right) -> Bool in
+            guard let leftDate = left.date,
+                let rightDate = right.date else {
+                    return false
             }
+            
+            return leftDate > rightDate
+        }
     }
 }
